Move unix FileSystem implementation to scalantive.nio.fs.unix, extract generic parts to GenericFileSystemProvider
