+++ javalib/src/main/scala/scala/scalanative/nio/fs/unix/UnixFileSystem.scala
@@ -1,4 +1,4 @@
-package scala.scalanative.nio.fs
+package scala.scalanative.nio.fs.unix
 
 import java.io.IOException
 import java.lang.Iterable
